- renaming typing typescript file - create chop interface - adding chop interface
(i) src/App.test.tsx
(i) src/App.tsx
(i) src/chop.tsx
(i) src/Typing/Types.ts -> src/Typing/Dataset.ts
(+)src/Typing/chop.ts
