+++ src/Typing/chop.ts
@@ -1,0 +1,4 @@
+export interface IChop {
+    data: any 
+    template: any
+}
